chore: rename jest-config package to avoid conflict
