+++ package.json
@@ -46,6 +46,7 @@
     "eslint-config-carbon": "2.2.0",
     "husky": "^4.3.6",
     "jest": "^25.5.4",
+    "jest-config-ibm-cloud-cognitive": "^0.3.11",
     "lerna": "^4.0.0",
     "lint-staged": "^11.0.0",
     "npm-check-updates": "^11.6.0",
