+++ packages/jest-config-ibm-cloud-cognitive/index.js
