+++ packages/jest-config-ibm-cloud-cognitive/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "jest-config",
+  "name": "jest-config-ibm-cloud-cognitive",
   "private": true,
   "version": "0.3.11",
   "license": "Apache-2.0",
@@ -7,7 +7,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/carbon-design-system/ibm-cloud-cognitive.git",
-    "directory": "packages/jest-config"
+    "directory": "packages/jest-config-ibm-cloud-cognitive"
   },
   "bugs": "https://github.com/carbon-design-system/ibm-cloud-cognitive/issues",
   "keywords": [
