+++ packages/jest-config-ibm-cloud-cognitive/setup/setupFilesAfterEnv.js
