+++ packages/jest-config-ibm-cloud-cognitive/transform/file.js
