+++ packages/jest-config-ibm-cloud-cognitive/transform/javascript.js
