+++ packages/security/package.json
@@ -59,7 +59,7 @@
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.6",
     "jest": "^25.5.4",
-    "jest-config": "^0.3.11",
+    "jest-config-ibm-cloud-cognitive": "^0.3.11",
     "npm-run-all": "^4.1.5"
   }
 }
